requirements.txt
pandas==2.2.3
numpy==2.2.3
pyyaml==6.0.2
click==8.1.7
starlette==0.46.1
uvicorn==0.34.0
filelock==3.18.0
tqdm==4.66.5
colorama==0.4.6
rpds_py==0.23.1
sgmllib3k==1.0.0
platformdirs==4.3.6
pluggy==1.5.0
primp==0.14.0
coverage==7.6.12
greenlet==3.1.1
h11==0.14.0
aiohttp==3.11.13
aiosignal==1.3.2
frozenlist==1.5.0
structlog==25.3.0
prompt-toolkit==3.0.50
wcwidth==0.2.13
six==1.17.0
langsmith==0.3.13
litellm==1.63.11
huggingface_hub==0.29.3
tiktoken==0.9.0
idna==3.10
importlib-metadata==8.5.0
jinja2==3.1.6
zipp==3.21.0
black==25.1.0
blockbuster==1.5.24
yarl==1.18.3
certifi==2025.1.31
distro==1.8.0
sniffio==1.3.1
socksio==1.0.0
tzdata==2025.1
urllib3==2.3.0
orjson==3.10.15
ormsgpack==1.9.1
webencodings==0.5.1
pathspec==0.12.1
mcp==1.6.0
sse-starlette
pyjwt==2.10.1
httpx==0.27.2
anyio==4.6.0
pydantic==2.9.2
pydantic-settings
httpx-sse
watchfiles==1.0.5
InquirerPy
langgraph
langchain_mcp_adapters
langchain_openai
markdownify
readabilipy
langchain_experimental
json_repair
loguru
fastapi
sentence-transformers
faiss-cpu
scikit-learn